+++ raytracer/features/Material/Material.h
@@ -13,10 +13,10 @@
 class Material {
 public:
     Color color = Color(1, 1, 1);
-    float ambient = 0.1;
-    float diffuse = 0.9;
-    float specular = 0.9;
-    float shininess = 200.0;
+    double ambient = 0.1;
+    double diffuse = 0.9;
+    double specular = 0.9;
+    double shininess = 200.0;
 
     bool operator==(const Material &other);
